+++ examples/map_events.rs
@@ -15,11 +15,6 @@ fn main() {
         // bevy_ecs_tilemap and bevy_ecs_tiled main plugins
         .add_plugins(TilemapPlugin)
         .add_plugins(TiledMapPlugin::default())
-        // Add observers for map loading events
-        .add_observer(map_created)
-        .add_observer(layer_created)
-        .add_observer(object_created)
-        .add_observer(tile_created)
         // Add our systems and run the app!
         .add_systems(Startup, startup)
         .run();
@@ -27,7 +22,13 @@ fn main() {
 
 fn startup(mut commands: Commands, asset_server: Res<AssetServer>) {
     commands.spawn(Camera2d);
-    commands.spawn(TiledMapHandle(asset_server.load("finite.tmx")));
+    commands
+        .spawn(TiledMapHandle(asset_server.load("finite.tmx")))
+        // Add observers for map loading events
+        .observe(map_created)
+        .observe(layer_created)
+        .observe(object_created)
+        .observe(tile_created);
 }
 
 fn map_created(