+++ src/world/events.rs
@@ -6,12 +6,17 @@ use crate::prelude::*;
 use bevy::prelude::*;
 
 /// Event sent when a Tiled world has finished loading.
-#[derive(Event, Clone, Debug)]
+#[derive(Component, Clone, Debug)]
 pub struct TiledWorldCreated {
     /// Spawned world [Entity].
     pub entity: Entity,
     /// [AssetId] of the corresponding [super::asset::TiledWorld] asset.
     pub asset_id: AssetId<TiledWorld>,
+}
+
+impl Event for TiledWorldCreated {
+    type Traversal = &'static Parent;
+    const AUTO_PROPAGATE: bool = true;
 }
 
 impl<'a> TiledWorldCreated {